[refactor] #14 상세 정보 뷰 관련 패키지 정리
