+++ app/src/main/AndroidManifest.xml
@@ -26,7 +26,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".presentation.DetailActivity"
+            android:name=".presentation.detail.DetailActivity"
             android:exported="false" />
         <activity
             android:name=".presentation.SearchResultActivity"
