+++ app/src/main/java/com/sopt/oliveyoung/application/OliveYoungApplication.kt
@@ -2,7 +2,7 @@ package com.sopt.oliveyoung.application
 
 import android.app.Application
 import com.sopt.oliveyoung.BuildConfig
-import com.sopt.oliveyoung.util.extensions.OliveYoungDebugTree
+import com.sopt.oliveyoung.util.OliveYoungDebugTree
 import timber.log.Timber
 
 class OliveYoungApplication : Application() {
