+++ app/src/main/res/layout/activity_detail.xml
@@ -7,13 +7,13 @@
 
         <variable
             name="viewModel"
-            type="com.sopt.oliveyoung.presentation.ProductDetailViewModel" />
+            type="com.sopt.oliveyoung.presentation.detail.ProductDetailViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".presentation.DetailActivity">
+        tools:context=".presentation.detail.DetailActivity">
 
         <androidx.constraintlayout.widget.ConstraintLayout
             android:id="@+id/layout_toolbar"
